backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
pypdf==3.17.1
python-docx==1.1.0
langchain==0.1.0
langchain-text-splitters==0.0.1
google-generativeai==0.3.2
pinecone-client==2.2.4
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4